- package.json: upgraded prisma and @prisma/client to ^7.8.0, added @prisma/adapter-pg + pg deps, added @types/pg + tsx devDeps, removed unused   @next-auth/prisma-adapter, added postinstall script   - prisma/schema.prisma: updated generator to prisma-client with output path, removed datasource URL   - prisma.config.ts: new file, holds datasource URL via env   - lib/db.js → lib/db.ts: rewritten to use PrismaPg adapter and import from generated path   - tsconfig.json / jsconfig.json: moduleResolution: bundler   - .gitignore: added /lib/generated   - Test mocks: updated imports from @prisma/client to lib/generated/prisma/client
