+++ package.json
@@ -23,11 +23,12 @@
     "migrate:dev": "npx prisma migrate dev",
     "migrate:reset": "npx prisma migrate reset",
     "heroku-prebuild": "npx prisma migrate deploy",
+    "postinstall": "prisma generate",
     "test": "jest --watch --silent false"
   },
   "dependencies": {
-    "@next-auth/prisma-adapter": "^1.0.7",
-    "@prisma/client": "^5.8.0",
+    "@prisma/adapter-pg": "^7.8.0",
+    "@prisma/client": "^7.8.0",
     "@sentry/nextjs": "^10.40.0",
     "ansi-colors": "^4.1.3",
 "cli-progress": "^3.12.0",
@@ -37,7 +38,8 @@
     "next": "^16.2.6",
     "next-auth": "^4.24.12",
     "next-i18next": "^15.2.0",
-    "prisma": "^5.8.0",
+    "pg": "^8.13.3",
+    "prisma": "^7.8.0",
     "react": "^19.2.3",
     "react-dom": "^19.2.3",
     "react-error-boundary": "^4.0.12",
@@ -57,6 +59,7 @@
     "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.5.1",
     "@types/fast-html-parser": "^1.0.4",
+    "@types/pg": "^8.11.0",
     "@types/jest": "^29.5.8",
     "@types/node": "^25.3.3",
     "@types/react": "^19.2.3",
@@ -81,6 +84,7 @@
     "stylelint-config-standard": "^40.0.0",
     "stylelint-order": "^7.0.1",
     "stylelint-scss": "^7.0.0",
+    "tsx": "^4.19.3",
     "typescript": "^5.3.3",
     "typescript-eslint": "^8.0.0"
   },
