+++ prisma.config.ts
@@ -1,0 +1,12 @@
+import 'dotenv/config';
+import { defineConfig, env } from 'prisma/config';
+
+export default defineConfig({
+  schema: 'prisma/schema.prisma',
+  migrations: {
+    path: 'prisma/migrations',
+  },
+  datasource: {
+    url: env('HEROKU_POSTGRESQL_BLUE_URL'),
+  },
+});
